+++ producao.h
@@ -1,6 +1,6 @@
 #ifndef PRODUCAO_H
 #define PRODUCAO_H
 
-void lp_solve_format(int n, int m, int* precos, int** compostos, float** quantidades);
+void lp_solve_format(int n, int m, double* precos, double** compostos, double** quantidades);
 
 #endif
